requirements.txt
amqp==1.4.9
anyjson==0.3.3
azure-cognitiveservices-language-textanalytics==0.2.0
azure-common==1.1.23
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-celery==3.3.1
django-kombu==0.9.4
djangorestframework==3.10.3
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
isodate==0.6.0
kombu==3.0.37
more-itertools==7.2.0
msrest==0.6.10
oauthlib==3.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-instagram==1.3.2
pytz==2019.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
slugify==0.0.1
sqlparse==0.3.0
urllib3==1.25.6
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
